test(user): split ManagerTest into DB and unit-only classes
The 28 pure-mock methods no longer carry #[Group('DB')] and will no
longer run across all 5 database backends. The 4 methods that create
real users and query the database stay in ManagerTest with #[Group('DB')].

AI-Assisted-By: Claude Sonnet 4.6 <noreply@anthropic.com>
